feat(osrm): roll out shared OSRM clients across delivery services
Move the OSRM spec and deployment assets into a dedicated root module so both services generate clients from one source of truth and return real snapped addresses in delivery flows.

Made-with: Cursor

+++ delivery/src/infrastructure/repository/courier_postgres.rs
@@ -259,7 +259,7 @@ impl CourierRepository for CourierPostgresRepository {
             .await
             .map_err(|e| RepositoryError::QueryError(e.to_string()))?;
 
-        Ok(count as u64)
+        Ok(count)
     }
 }
